Source/Services/RPSLS.PythonPlayer.Api/requirements.txt
gunicorn
flask
py-healthcheck